KAFKA-13461: Don't re-initialize ZK client session after auth failure if connection still alive (#11563)
Cherry-pick of da56146c60cf016966b119ea63eb2cfa60f03ad4

If JAAS configuration does not contain a Client section for ZK clients, an auth failure event is generated. If this occurs after the connection is setup in the controller, we schedule reinitialize(), which causes controller to resign. In the case where SASL is not mandatory and the connection is alive, controller maintains the current session and doesn't register its watchers, leaving it in a bad state.

Reviewers: Jun Rao <junrao@gmail.com>