DP-9030: Use the new withGradleFile closure (#842)
Confluent is migrating from JFrog to CodeArtifact. The current vault secret has the contents of a `gradle.init`
file with credentials for JFrog. CodeArtifact's credentials only last 12 hours, so it needs to be refreshed.

The `withGradleFile` closure wraps the `withVaultFile` closure, but refreshes and substitutes to CodeArtifact
token into the file.

Merging this pull request **will not** cause the repo to publish to CodeArtifact. When it is time to migrate
(later this week), we will swap out the underlying vault secret.

https://confluentinc.atlassian.net/wiki/spaces/TOOLS/pages/2930704487/Java+Migration+FAQ#Jenkins

I tested this new closure in Jenkins with a custom job.

Reviewers: Ismael Juma <ismael@juma.me.uk>, Adam Brown <acourtneybrown@users.noreply.github.com>, Andy Lapidas <alapidas@gmail.com>