Revert "MINOR: Fix some compiler warnings (#12912)"
This reverts commit 42cfd57a2498b1a997e77cc24ec38b3722cc1f6a.

The right branching point should have been 96b1db510a32b99.
I incorrectly picked the branching point to be the commit before
the version bump, but there was a gap between the branching
point and the version bump merge in AK trunk.

Most of the additional commits were cherry-picked to 3.4, but
this one was not.