Revert "KAFKA-14367; Add `ListGroups` to the new `GroupCoordinator` interface (#12853)"
This reverts commit 854dfb5ffcec279d4df003374df8984296d264f1.

The right branching point should have been 96b1db5.
I incorrectly picked the branching point to be the commit before
the version bump, but there was a gap between the branching
point and the version bump merge in AK trunk.

Most of the additional commits were cherry-picked to 3.4, but
this one was not.